requirements.txt
assemblyai
python-dotenv
anthropic